HYPERRSI/requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioschedule==0.5.2
aiosignal==1.3.2
aiosqlite==0.20.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==24.3.0
billiard==4.2.1
binance==0.3
ccxt==4.4.50
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
deprecation==2.1.0
fastapi==0.115.6
fonttools==4.55.3
frozenlist==1.5.0
gotrue==2.11.1
greenlet==3.1.1
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
kiwisolver==1.4.8
kombu==5.4.2
magic-filter==1.0.12
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
postgrest==0.19.1
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.10
pycares==4.5.0
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.5
rq==2.1.0
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.5.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
wheel==0.45.1
yarl==1.18.3
aiodns==3.2.0
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioschedule==0.5.2
aiosignal==1.3.2
aiosqlite==0.20.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==24.3.0
billiard==4.2.1
binance==0.3
ccxt==4.4.50
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
deprecation==2.1.0
fastapi==0.115.6
flower==2.0.1
fonttools==4.55.3
frozenlist==1.5.0
gotrue==2.11.1
greenlet==3.1.1
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
kombu==5.4.2
magic-filter==1.0.12
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
postgrest==0.19.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.10
pycares==4.5.0
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.1
PySide6==6.8.1.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.5
rq==2.1.0
scipy==1.15.1
setuptools==75.8.0
shiboken6==6.8.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
tenacity==9.0.0
tornado==6.5.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.5.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
wheel==0.45.1
yarl==1.18.3
